Add all contents into randomplayer.py
